Reorganizing the common code for aggregate functions
