+++ operations/common/go.mod
@@ -1,3 +1,5 @@
 module github.com/skothari-tibco/operations-sample/common
 
 go 1.12
+
+require github.com/project-flogo/core v0.9.3
